replace new_fp
