--- crates/new2-fp/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "new2_fp"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
