+++ crates/new_fp/Cargo.toml
@@ -1,11 +1,9 @@
 [package]
 name = "new_fp"
 version = "0.1.0"
+authors = ["Kana <ngtkana@gmail.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-
-[dev-dependencies]
-assert-impl = "0.1.3"
